repair6

+++ src/main/java/com/paraciuman/hackathon/repository/AgendaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
 @Component
-public interface AgendaRepository extends JpaRepository<Agenda, Integer> {
+public interface AgendaRepository extends JpaRepository<Agenda, Long> {
 }
