+++ src/main/java/com/paraciuman/hackathon/repository/CheckListRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
 @Component
-public interface CheckListRepository extends JpaRepository<CheckList, Integer> {
+public interface CheckListRepository extends JpaRepository<CheckList, Long> {
 }
